+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - build(android): Changes android package name from `io.sentry.RNSentryPackage` to `io.sentry.react.RNSentryPackage` (Breaking). #1131
 - fix: As auto session tracking is now on by default, allow user to pass `false` to disable it. #1131
-- build: Bump `sentry-android` to 3.0.0. #1131
+- build: Bump `sentry-android` to 3.1.0. #1131
 - build: Bump `sentry-cocoa` to 6.0.3. #1131
 - feat(ios): Use `captureEnvelope` on iOS/Mac. #1131
 - feat: Support envelopes with type other than `event`. #1131