+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.0.0
+
 - build(android): Changes android package name from `io.sentry.RNSentryPackage` to `io.sentry.react.RNSentryPackage` (Breaking). #1131
 - fix: As auto session tracking is now on by default, allow user to pass `false` to disable it. #1131
 - build: Bump `sentry-android` to 3.1.0. #1131